refactor : 불필요한 주석 및 공백 삭제
