+++ src/components/CartBadge/CartBadge.tsx
@@ -1,10 +1,9 @@
-import { useCart } from '../../context/ShoppingCartCountContext';
+import { useCart } from '../../context/CartContext';
 import Badge from '../common/Badge/Badge';
 import { CartBadgeContainer } from './CartBadge.style';
 import { PropsWithChildren } from 'react';
 
 const CartBadge: React.FC<PropsWithChildren> = ({ children }) => {
-  // const [counts, setCounts] = useState<number>(0);
   const { counts } = useCart();
 
   return (
